делит код на модули (FSD-архитектура)
